DataBuilder

Documentation

getData :: Filepath -> Separator -> IO DataContainer

Takes a path to the file, seperator that seperates data and returns DataContainer

filterDataContainer :: [Int] -> DataContainer -> NumContainer

Takes list of indexes of columns that are to be deleted, DataContainer with data and return NumContainer with only numeric values. | If it cannot get a number out of a field it removes this one training set.

getMatrixFromDataContainer :: DataContainer -> Matrix String

getMatrixFromNumContainer :: NumContainer -> Matrix Double

newtype NumContainer

NumContainer stores the same as DataContainer except non-numerical values

Constructors

NumContainer ([Header], Matrix Double)

type Header = String

data DataContainer

DataContainer is a container for headers of data and data itself