Matrix

Description

Module Matrix provides simple operatons on 2d matrixes and matrix datatype as well.

Matrix is represented as list of lists. Each list in the list is a column.

Documentation

data Matrix a

This is Matrix, represented as list of lists in a package

emptyM :: Matrix a

Gives an empty matrix

vector :: [a] -> Axis -> Matrix a

Gives a vector

zipWithM :: (a -> b -> c) -> Matrix a -> Matrix b -> MaybeRes (Matrix c)

Zips corresponding elements using function

transposeM :: Matrix a -> Matrix a

Transposes matrix

howManyElements :: Matrix a -> Int

Gives a number of elements in matrix

toRectMatrix :: Width -> Height -> [a] -> Matrix a

Gives a matrix of sizes specified in arguments.

findEdgeSizes :: [a] -> [(Width, Height)]

Gives pairs of sizes as tuples (Width, Height) in which matrix can be represented

findAMostQuadraticEdgeSize :: [a] -> (Width, Height)

Gives a pair of sizes in a tuple (Width, Height) most close to square in which matrix can be represented

toASquareMatrix :: [a] -> Matrix a

Gives a matrix of sizes most close to square

toHeightMatrix :: Height -> [a] -> Matrix a

Gives a matrix of height Height and width counted so that it stays rectangular

toList :: Matrix a -> [a]

Return flat list of elements of matrix

normalize :: Matrix a -> Matrix a

Makes matrix as quadratic as possible

zipWithLines :: (a -> a -> a) -> Matrix a -> Matrix a

Zips columns using specified function.

conver :: Matrix a -> Matrix a -> MaybeRes (Matrix a)

Concatenates matrixes in vertical axis. Possible infix notation

conhor :: Matrix a -> Matrix a -> MaybeRes (Matrix a)

Concatenates matrixes in horizontal axis. Possible infix notation

filterLinesHor :: (a -> Bool) -> Matrix a -> Matrix a

Checks if all of elements in horizontal line fulfill predicate, leaves those lines that fulfill

filterLinesVer :: (a -> Bool) -> Matrix a -> Matrix a

Checks if all of elements in vertical line fulfill predicate, leaves those lines that fulfill

deleteColumns :: [Int] -> Matrix a -> Matrix a

Deletes all the columns specified as their indexes in the list given as argument, indexing from 1

scaleLines :: (Num a, Fractional a, Real a) => Matrix a -> Matrix a

Scales columns, it substracts the average of column and devides by range of elements in a column for each element of a column.

getColumn :: Int -> Matrix a -> Matrix a

Returns n-th column as a vertical vector, indexing from 1

getRow :: Int -> Matrix a -> Matrix a

Returns n-th row as a horizontal vector, indexing from 1

unpackM :: Matrix a -> [[a]]

Gives a list of lists out of matrix

packM :: [[a]] -> Matrix a

Packs a list of lists into a matrix

getElementByInd :: (Width, Height) -> Matrix a -> MaybeRes a

Returns element that is on index (Width, Height), indexing from 1

getSize :: Matrix a -> (Width, Height)

Returns size of matrix as tuple (Width, Height)

getWidth :: Matrix a -> Width

Returns Width of matrix

getHeight :: Matrix a -> Height

Returns Height of matrix

getAvgColumnsMatrix :: (Num a, Fractional a, Real a) => Matrix a -> Matrix a

Retruns a horizontal vector of averages for each column

getRangeColumnsMatrix :: (Num a, Fractional a, Real a) => Matrix a -> Matrix a

Retruns a horizontal vector of ranges for each column

isEmpty :: Matrix a -> Bool

Returns true if Matrix is Empty