Regressions

Documentation

scale :: XYContainer -> MaybeRes XYContainer

Scales data in XYContainer. For each element: x' = (x - avg(column where x is)) / range_of_values(column where x is)

addFreeTerm :: XYContainer -> MaybeRes XYContainer

Adds a column of ones at the end of X.

getXYContainer :: Int -> NumContainer -> XYContainer

Takes an y column index, NumContainer and produces XYContainer - holder of X matrix, its headers, Y vector, information | about amount of basic features and information about if free term column has been added to X matrix. Indexing from 1

getX :: XYContainer -> X

getY :: XYContainer -> Y

getHeaders :: XYContainer -> [Header]

getAmountOfBasicFeatures :: XYContainer -> AmountOfBasicFeatures

gradientDescent :: XYContainer -> Alpha -> Epsilon -> MaxIter -> [(NormalizeConst, VariableInd)] -> IO GradDescentInfo

Performs a gradient descent algorithm for data in XYContainer. Alpha is a learn speed rate, epsilon is the convergence | condition (gradient descent stops when the difference between current cost and next cost is lower than epsilon), | MaxIter is a maximum of iteration you allow gradient descent to go through, and a list of (NormalizeConst, VariableInd) | is a list pairs of normalize const and variable index - it adds to the cost function the sum of products of elements of each pair.

getLastTheta :: GradDescentInfo -> Theta

getLastCost :: GradDescentInfo -> Cost

makeTrainingSet :: [Double] -> TrainingSet

Creates a TrainingSet out of a list

createNewFeature :: VariableInd -> Power -> XYContainer -> MaybeRes XYContainer

Creates new features to X by taking VariableInd-th feature and raise it to the power Power

createNewFeaturesFromListToTrainingSet :: [NewFeature] -> TrainingSet -> TrainingSet

Creates new features to the training set as specified in a list of info about new features. XYContainer is a base | container - before scaling and adding free term to it

evaluateInputData :: TrainingSet -> Theta -> XYContainer -> XYContainer -> Double

Evaluates input data (just basic features needed to pass), Param: input data, theta, baseContainer (before scaling), fullyDoneContainer (after adding all the staff)

type GradDescentInfo = (Theta, Cost)

Holds information about training. As a tuple: (Produced Theta, Last Cost)

type Alpha = Double

Learn speed rate

type Epsilon = Double

Convergence condition. Specifies how small should the difference between next costs be to stop gradient descent.

type Cost = Double

Cost of polynomial fitting

type Theta = Matrix Double

type NormalizeConst = Double

Const that is used to multiply one of thetas element and then add the product to a cost function

type TrainingSet = Matrix Double

Horizontal vector of data of one single training set

data XYContainer

holder of X matrix, its headers, Y vector, information | about amount of basic features and information about if free term column has been added to X matrix, | a list of new features (tuples with information to what power which variable has been raised to in order to create new feature)

type VariableInd = Int

Index of a variable

type NewFeature = (Power, VariableInd)

Tuple with information on an added feature to data, in detail: about a power | to which variable on index VariableInd has been raised in order to create new feature